+++ src/experimentation-opportunities/experimentation-opportunities.js
@@ -167,6 +167,7 @@ async function updateRecommendations(oppty, context, site) {
   let lambdaResult;
   try {
     // eslint-disable-next-line no-await-in-loop
+    log.info('Invoking lambda function with payload: ', JSON.stringify(lambdaPayload, null, 2));
     const lambdaResponse = await invokeLambdaFunction(lambdaPayload);
     log.info('Lambda Response: ', JSON.stringify(lambdaResponse, null, 2));
     const lambdaResponseBody = typeof lambdaResponse.body === 'string'
@@ -322,9 +323,10 @@ function convertToOpportunityEntity(oppty, auditData) {
 export async function postProcessor(auditUrl, auditData, context) {
   const { log } = context;
   const { dataAccess } = context;
+  const { Opportunity } = dataAccess;
   let updatedEntities = 0;
   log.info(`Experimentation Opportunities post processing started for ${auditUrl} from audit ${auditData.id}`);
-  const existingOpportunities = await dataAccess.Opportunity.allBySiteId(auditData.siteId);
+  const existingOpportunities = await Opportunity.allBySiteId(auditData.siteId);
 
   // Get opportunities with recommendations
   const opportunities = auditData.auditResult.experimentationOpportunities
@@ -370,7 +372,6 @@ export async function handler(auditUrl, context, site) {
     interval: DAYS,
     granularity: 'hourly',
   };
-
   const queryResults = await rumAPIClient.queryMulti(OPPTY_QUERIES, options);
   const experimentationOpportunities = Object.values(queryResults).flatMap((oppty) => oppty);
   await processHighOrganicLowCtrOpportunities(experimentationOpportunities, context, site);