+++ test/audits/metatags.test.js
@@ -194,11 +194,18 @@ describe('Meta Tags', () => {
       sinon.restore();
       logStub = { info: sinon.stub(), error: sinon.stub(), warn: sinon.stub() };
       dataAccessStub = {
-        getConfiguration: sinon.stub(),
-        getTopPagesForSite: sinon.stub(),
-        addAudit: sinon.stub(),
-        retrieveSiteBySiteId: sinon.stub(),
-        getSiteByID: sinon.stub().resolves({ isLive: sinon.stub().returns(true) }),
+        Audit: {
+          create: sinon.stub(),
+        },
+        Configuration: {
+          findLatest: sinon.stub(),
+        },
+        Site: {
+          findById: sinon.stub().resolves({ getIsLive: sinon.stub().returns(true) }),
+        },
+        SiteTopPage: {
+          allBySiteId: sinon.stub(),
+        },
       };
       s3ClientStub = {
         send: sinon.stub(),
@@ -263,7 +270,7 @@ describe('Meta Tags', () => {
           ContentType: 'application/json',
         });
       const addAuditStub = sinon.stub().resolves({ getId: () => 'audit-id' });
-      dataAccessStub.addAudit = addAuditStub;
+      dataAccessStub.Audit.create = addAuditStub;
 
       expect(addAuditStub.calledWithMatch({
         '/blog/page1': {
@@ -323,7 +330,7 @@ describe('Meta Tags', () => {
 
     xit('should process site tags and perform SEO checks for pages with invalid H1s', async () => {
       const site = {
-        isLive: sinon.stub().returns(true),
+        getIsLive: sinon.stub().returns(true),
         getId: sinon.stub().returns('site-id'),
         getBaseURL: sinon.stub().returns('http://example.com'),
       };
@@ -331,11 +338,11 @@ describe('Meta Tags', () => {
         { getURL: 'http://example.com/blog/page2', getTopKeyword: sinon.stub().returns('Test') },
         { getURL: 'http://example.com/', getTopKeyword: sinon.stub().returns('Test') }];
 
-      dataAccessStub.getSiteByID.resolves(site);
-      dataAccessStub.getConfiguration.resolves({
+      dataAccessStub.Site.findById.resolves(site);
+      dataAccessStub.Configuration.findLatest.resolves({
         isHandlerEnabledForSite: sinon.stub().returns(true),
       });
-      dataAccessStub.getTopPagesForSite.resolves(topPages);
+      dataAccessStub.SiteTopPage.allBySiteId.resolves(topPages);
 
       s3ClientStub.send
         .withArgs(sinon.match.instanceOf(ListObjectsV2Command).and(sinon.match.has('input', {
@@ -411,7 +418,7 @@ describe('Meta Tags', () => {
           ContentType: 'application/json',
         });
       const addAuditStub = sinon.stub().resolves();
-      dataAccessStub.addAudit = addAuditStub;
+      dataAccessStub.Audit.create = addAuditStub;
 
       expect(addAuditStub.calledWithMatch({
         '/blog/page1': {
@@ -486,11 +493,18 @@ describe('Meta Tags', () => {
     });
 
     it('should handle gracefully if S3 object has no rawbody', async () => {
+      // const site = {
+      //   getIsLive: sinon.stub().returns(true),
+      //   getId: sinon.stub().returns('site-id'),
+      //   getBaseURL: sinon.stub().returns('http://example.com'),
+      // };
       const topPages = [{ getURL: 'http://example.com/blog/page1', getTopKeyword: sinon.stub().returns('page') }];
-      dataAccessStub.getConfiguration.resolves({
+
+      // dataAccessStub.Site.findById.resolves(site);
+      dataAccessStub.Configuration.findLatest.resolves({
         isHandlerEnabledForSite: sinon.stub().returns(true),
       });
-      dataAccessStub.getTopPagesForSite.resolves(topPages);
+      dataAccessStub.SiteTopPage.allBySiteId.resolves(topPages);
 
       s3ClientStub.send
         .withArgs(sinon.match.instanceOf(ListObjectsV2Command).and(sinon.match.has('input', {
@@ -522,18 +536,18 @@ describe('Meta Tags', () => {
 
     it('should handle gracefully if S3 tags object is not valid', async () => {
       const site = {
-        isLive: sinon.stub().returns(true),
+        getIsLive: sinon.stub().returns(true),
         getId: sinon.stub().returns('site-id'),
         getBaseURL: sinon.stub().returns('http://example.com'),
       };
       const topPages = [{ getURL: 'http://example.com/blog/page1', getTopKeyword: sinon.stub().returns('page') },
         { getURL: 'http://example.com/blog/page2', getTopKeyword: sinon.stub().returns('Test') }];
 
-      dataAccessStub.getSiteByID.resolves(site);
-      dataAccessStub.getConfiguration.resolves({
+      dataAccessStub.Site.findById.resolves(site);
+      dataAccessStub.Configuration.findLatest.resolves({
         isHandlerEnabledForSite: sinon.stub().returns(true),
       });
-      dataAccessStub.getTopPagesForSite.resolves(topPages);
+      dataAccessStub.SiteTopPage.allBySiteId.resolves(topPages);
 
       s3ClientStub.send
         .withArgs(sinon.match.instanceOf(ListObjectsV2Command).and(sinon.match.has('input', {